Verification Link Success
The verfication link consist of a uid and a unique time dependent token.
uid is the userid which is encoded in urlbase64 with force_text so that
it will be displayed easily in url in encoded form. Also token was
created from passwordresettoken generator with user.id, user.is_active,
and time stamp. Finally a url was created and on visiting url we
decoded uid with force_str and urlbase64decode and token with checktoken
. When all was right we activated the user and redirected to login.