Orders, OrderItems Completed
When user clicks pay, the respective orders and orderitems will be
created in the database and paymentconfirmation will be sent to
stripe, when orders are added to database, the related session are
also deleted. Now, the next problem is When payment is successfull the
order will have the billing status set to true, with the use of webhooks.