STRIPE HOOKS Done
The order_key in orders was initially the value of client_secret
from payment intent, the problem however was that the clinet_secret
was not send from the post data in stripe hooks, so we changed the
value of order_key from secret to payment_intent which is sent during
post request from stripe hooks. So, both clinet_secret, and payment_intent
had to be sent to templates because a payment cannot be made in stripe
without the client_secret. We have to build the orders placed in dashboard.