Fixed LoginRequiredBug, Added secret to orders
We need client_secret to initiate a payment so when orders are saved for
later, we cannot complete that payment without secret, so it is Added to
order model. Similarly, saving orders without login doesn't work which is
excepted behaviour, but error was shown since the response was html instead
of json due to the mixin, so it was fixed with catch in fetch.