Saved Order Payment
The saved orders shown in dashboard can now be paid with pay now button.
A new View was created that takes parameter intent from url, and then
queries the order db for secret, which is passed with session and then
sent to checkoutView that processes payment. There is an extra query to
db which will be refactored later.